requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==4.0.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
python-dotenv==1.0.0
pymongo==4.6.1
